+++ UC4-chaincode/src/main/java/de/upb/cs/uc4/chaincode/contract/certificate/CertificateContract.java
@@ -3,6 +3,7 @@ package de.upb.cs.uc4.chaincode.contract.certificate;
 import de.upb.cs.uc4.chaincode.contract.ContractBase;
 import de.upb.cs.uc4.chaincode.exceptions.SerializableError;
 import de.upb.cs.uc4.chaincode.exceptions.serializable.ParameterError;
+import de.upb.cs.uc4.chaincode.helper.HyperledgerManager;
 import org.hyperledger.fabric.contract.Context;
 import org.hyperledger.fabric.contract.annotation.Contract;
 import org.hyperledger.fabric.contract.annotation.Transaction;
@@ -30,7 +31,8 @@ public class CertificateContract extends ContractBase {
      */
     @Transaction()
     public String addCertificate(final Context ctx, final String enrollmentId, final String certificate) {
-        String transactionName = ctx.getStub().getTxId().split(":")[1];
+        String transactionName = HyperledgerManager.getTransactionName(ctx.getStub());
+
         try {
             cUtil.checkParamsAddCertificate(ctx, new ArrayList<String>(){{add(enrollmentId); add(certificate);}});
         } catch (ParameterError e) {
@@ -39,12 +41,12 @@ public class CertificateContract extends ContractBase {
 
         ChaincodeStub stub = ctx.getStub();
         try {
-            cUtil.validateApprovals(stub, this.contractName,  transactionName, new ArrayList<String>(){{add(enrollmentId);add(certificate);}});
+            cUtil.validateApprovals(ctx, contractName,  transactionName, new String[]{enrollmentId, certificate});
         } catch (SerializableError e) {
             return e.getJsonError();
         }
         try {
-            cUtil.finishOperation(stub, this.contractName,  transactionName, new ArrayList<String>(){{add(enrollmentId);add(certificate);}});
+            cUtil.finishOperation(stub, contractName,  transactionName, new String[]{enrollmentId, certificate});
         } catch (SerializableError e) {
             return e.getJsonError();
         }
@@ -61,7 +63,7 @@ public class CertificateContract extends ContractBase {
      */
     @Transaction()
     public String updateCertificate(final Context ctx, final String enrollmentId, final String certificate) {
-        String transactionName = ctx.getStub().getTxId().split(":")[1];
+        String transactionName = HyperledgerManager.getTransactionName(ctx.getStub());
         try {
             cUtil.checkParamsUpdateCertificate(ctx, new ArrayList<String>(){{add(enrollmentId); add(certificate);}});
         } catch (ParameterError e) {
@@ -70,12 +72,12 @@ public class CertificateContract extends ContractBase {
 
         ChaincodeStub stub = ctx.getStub();
         try {
-            cUtil.validateApprovals(stub, this.contractName,  transactionName, new ArrayList<String>(){{add(enrollmentId);add(certificate);}});
+            cUtil.validateApprovals(ctx, contractName,  transactionName, new String[]{enrollmentId, certificate});
         } catch (SerializableError e) {
             return e.getJsonError();
         }
         try {
-            cUtil.finishOperation(stub, this.contractName,  transactionName, new ArrayList<String>(){{add(enrollmentId);add(certificate);}});
+            cUtil.finishOperation(stub, contractName,  transactionName, new String[]{enrollmentId, certificate});
         } catch (SerializableError e) {
             return e.getJsonError();
         }
@@ -91,7 +93,7 @@ public class CertificateContract extends ContractBase {
      */
     @Transaction()
     public String getCertificate(final Context ctx, final String enrollmentId) {
-        String transactionName = ctx.getStub().getTxId().split(":")[1];
+        String transactionName = HyperledgerManager.getTransactionName(ctx.getStub());
         try {
             cUtil.checkParamsGetCertificate(ctx, Collections.singletonList(enrollmentId));
         } catch (ParameterError e) {
@@ -100,12 +102,12 @@ public class CertificateContract extends ContractBase {
 
         ChaincodeStub stub = ctx.getStub();
         try {
-            cUtil.validateApprovals(stub, this.contractName,  transactionName, Collections.singletonList(enrollmentId));
+            cUtil.validateApprovals(ctx, contractName,  transactionName, new String[]{enrollmentId});
         } catch (SerializableError e) {
             return e.getJsonError();
         }
         try {
-            cUtil.finishOperation(stub, this.contractName,  transactionName, Collections.singletonList(enrollmentId));
+            cUtil.finishOperation(stub, contractName,  transactionName, new String[]{enrollmentId});
         } catch (SerializableError e) {
             return e.getJsonError();
         }