+++ UC4-chaincode/src/main/java/de/upb/cs/uc4/chaincode/contract/group/GroupContract.java
@@ -4,6 +4,7 @@ import de.upb.cs.uc4.chaincode.contract.ContractBase;
 import de.upb.cs.uc4.chaincode.exceptions.serializable.LedgerAccessError;
 import de.upb.cs.uc4.chaincode.exceptions.serializable.ParameterError;
 import de.upb.cs.uc4.chaincode.exceptions.SerializableError;
+import de.upb.cs.uc4.chaincode.helper.HyperledgerManager;
 import de.upb.cs.uc4.chaincode.model.Group;
 import de.upb.cs.uc4.chaincode.helper.GsonWrapper;
 import org.hyperledger.fabric.contract.Context;
@@ -33,7 +34,7 @@ public class GroupContract extends ContractBase {
      */
     @Transaction()
     public String addUserToGroup(final Context ctx, String enrollmentId, String groupId) {
-        String transactionName = ctx.getStub().getTxId().split(":")[1];
+        String transactionName = HyperledgerManager.getTransactionName(ctx.getStub());
         try {
             cUtil.checkParamsAddUserToGroup(ctx, new ArrayList<String>(){{add(enrollmentId); add(groupId);}});
         } catch (ParameterError e) {
@@ -54,14 +55,14 @@ public class GroupContract extends ContractBase {
         }
 
         try {
-            cUtil.validateApprovals(stub, this.contractName, transactionName, new ArrayList<String>() {{add(enrollmentId);add(groupId);}});
+            cUtil.validateApprovals(ctx, contractName, transactionName, new String[]{enrollmentId, groupId});
         } catch (SerializableError e) {
             return e.getJsonError();
         }
 
         cUtil.putAndGetStringState(stub, groupId, GsonWrapper.toJson(group));
         try {
-            cUtil.finishOperation(stub, this.contractName, transactionName, new ArrayList<String>() {{add(enrollmentId);add(groupId);}});
+            cUtil.finishOperation(stub, contractName, transactionName, new String[]{enrollmentId, groupId});
         } catch (SerializableError e) {
             return e.getJsonError();
         }
@@ -78,7 +79,7 @@ public class GroupContract extends ContractBase {
      */
     @Transaction()
     public String removeUserFromGroup(final Context ctx, String enrollmentId, String groupId) {
-        String transactionName = ctx.getStub().getTxId().split(":")[1];
+        String transactionName = HyperledgerManager.getTransactionName(ctx.getStub());
         try {
             cUtil.checkParamsRemoveUserFromGroup(ctx, new ArrayList<String>(){{add(enrollmentId); add(groupId);}});
         } catch (SerializableError e) {
@@ -95,14 +96,14 @@ public class GroupContract extends ContractBase {
         group.getUserList().remove(enrollmentId);
 
         try {
-            cUtil.validateApprovals(stub, this.contractName, transactionName, new ArrayList<String>() {{add(enrollmentId);add(groupId);}});
+            cUtil.validateApprovals(ctx, contractName, transactionName, new String[]{enrollmentId, groupId});
         } catch (SerializableError e) {
             return e.getJsonError();
         }
 
         cUtil.putAndGetStringState(stub, groupId, GsonWrapper.toJson(group));
         try {
-            cUtil.finishOperation(stub, this.contractName, transactionName, new ArrayList<String>() {{add(enrollmentId);add(groupId);}});
+            cUtil.finishOperation(stub, contractName, transactionName, new String[]{enrollmentId, groupId});
         } catch (SerializableError e) {
             return e.getJsonError();
         }
@@ -118,7 +119,7 @@ public class GroupContract extends ContractBase {
      */
     @Transaction()
     public String removeUserFromAllGroups(final Context ctx, String enrollmentId) {
-        String transactionName = ctx.getStub().getTxId().split(":")[1];
+        String transactionName = HyperledgerManager.getTransactionName(ctx.getStub());
         try {
             cUtil.checkParamsRemoveUserFromAllGroups(Collections.singletonList(enrollmentId));
         } catch (ParameterError e) {
@@ -127,17 +128,17 @@ public class GroupContract extends ContractBase {
 
         ChaincodeStub stub = ctx.getStub();
         try {
-            cUtil.validateApprovals(stub, this.contractName, transactionName, Collections.singletonList(enrollmentId));
+            cUtil.validateApprovals(ctx, contractName, transactionName, new String[]{enrollmentId});
         } catch (SerializableError e) {
             return e.getJsonError();
         }
 
-        cUtil.getGroupsForUser(stub, enrollmentId).forEach(item -> {
-            item.getUserList().remove(enrollmentId);
-            cUtil.putAndGetStringState(stub, item.getGroupId(), GsonWrapper.toJson(item));
+        cUtil.getGroupsForUser(stub, enrollmentId).forEach(group -> {
+            group.getUserList().remove(enrollmentId);
+            cUtil.putAndGetStringState(stub, group.getGroupId(), GsonWrapper.toJson(group));
         });
         try {
-            cUtil.finishOperation(stub, this.contractName, transactionName, Collections.singletonList(enrollmentId));
+            cUtil.finishOperation(stub, contractName, transactionName, new String[]{enrollmentId});
         } catch (SerializableError e) {
             return e.getJsonError();
         }
@@ -152,16 +153,16 @@ public class GroupContract extends ContractBase {
      */
     @Transaction()
     public String getAllGroups(final Context ctx) {
-        String transactionName = ctx.getStub().getTxId().split(":")[1];
+        String transactionName = HyperledgerManager.getTransactionName(ctx.getStub());
         ChaincodeStub stub = ctx.getStub();
         try {
-            cUtil.validateApprovals(stub, this.contractName, transactionName, new ArrayList<String>());
+            cUtil.validateApprovals(ctx, contractName, transactionName, new String[]{});
         } catch (SerializableError e) {
             return e.getJsonError();
         }
         List<Group> groupList = cUtil.getAllGroups(stub);
         try {
-            cUtil.finishOperation(stub, this.contractName, transactionName, new ArrayList<String>());
+            cUtil.finishOperation(stub, contractName, transactionName, new String[]{});
         } catch (SerializableError e) {
             return e.getJsonError();
         }
@@ -176,7 +177,7 @@ public class GroupContract extends ContractBase {
      */
     @Transaction()
     public String getUsersForGroup(final Context ctx, String groupId) {
-        String transactionName = ctx.getStub().getTxId().split(":")[1];
+        String transactionName = HyperledgerManager.getTransactionName(ctx.getStub());
         try {
             cUtil.checkParamsGetUsersForGroup(ctx, Collections.singletonList(groupId));
         } catch (SerializableError e) {
@@ -185,7 +186,7 @@ public class GroupContract extends ContractBase {
 
         ChaincodeStub stub = ctx.getStub();
         try {
-            cUtil.validateApprovals(stub, this.contractName, transactionName, Collections.singletonList(groupId));
+            cUtil.validateApprovals(ctx, contractName, transactionName, new String[]{groupId});
         } catch (SerializableError e) {
             return e.getJsonError();
         }
@@ -196,7 +197,7 @@ public class GroupContract extends ContractBase {
             return e.getJsonError();
         }
         try {
-            cUtil.finishOperation(stub, this.contractName, transactionName, Collections.singletonList(groupId));
+            cUtil.finishOperation(stub, contractName, transactionName, new String[]{groupId});
         } catch (SerializableError e) {
             return e.getJsonError();
         }
@@ -212,7 +213,7 @@ public class GroupContract extends ContractBase {
      */
     @Transaction()
     public String getGroupsForUser(final Context ctx, String enrollmentId) {
-        String transactionName = ctx.getStub().getTxId().split(":")[1];
+        String transactionName = HyperledgerManager.getTransactionName(ctx.getStub());
         try {
             cUtil.checkParamsGetGroupsForUser(Collections.singletonList(enrollmentId));
         } catch (ParameterError e) {
@@ -221,13 +222,13 @@ public class GroupContract extends ContractBase {
 
         ChaincodeStub stub = ctx.getStub();
         try {
-            cUtil.validateApprovals(stub, this.contractName, transactionName, Collections.singletonList(enrollmentId));
+            cUtil.validateApprovals(ctx, contractName, transactionName, new String[]{enrollmentId});
         } catch (SerializableError e) {
             return e.getJsonError();
         }
         List<String> groupIdList = cUtil.getGroupNamesForUser(stub, enrollmentId);
         try {
-            cUtil.finishOperation(stub, this.contractName, transactionName, Collections.singletonList(enrollmentId));
+            cUtil.finishOperation(stub, contractName, transactionName, new String[]{enrollmentId});
         } catch (SerializableError e) {
             return e.getJsonError();
         }