+++ UC4-chaincode/src/main/java/de/upb/cs/uc4/chaincode/contract/operation/OperationContract.java
@@ -14,7 +14,6 @@ import org.hyperledger.fabric.contract.Context;
 import org.hyperledger.fabric.contract.annotation.Contract;
 import org.hyperledger.fabric.contract.annotation.Transaction;
 
-import java.lang.reflect.Array;
 import java.security.NoSuchAlgorithmException;
 import java.time.ZoneId;
 import java.time.ZoneOffset;
@@ -38,8 +37,13 @@ public class OperationContract extends ContractBase {
      * @return certificate on success, serialized error on failure
      */
     @Transaction()
-    public String approveTransaction(final Context ctx, final String initiator, final String contractName, final String transactionName, final String params) {
+    public String approveTransaction(final Context ctx, String initiator, final String contractName, final String transactionName, final String params) {
         ArrayList<InvalidParameter> invalidParameters = cUtil.getErrorForInput(contractName, transactionName);
+        String clientId = cUtil.getEnrollmentIdFromClientId(ctx.getClientIdentity().getId());
+        List<String> clientGroups = new GroupContractUtil().getGroupNamesForUser(ctx.getStub(), clientId);
+
+        initiator = initiator.isEmpty() ? clientId : initiator;
+
         if(!invalidParameters.isEmpty()){
             return GsonWrapper.toJson(cUtil.getUnprocessableEntityError(invalidParameters));
         }
@@ -70,11 +74,9 @@ public class OperationContract extends ContractBase {
                     .reason("");
 
         }
-        String clientId = cUtil.getEnrollmentIdFromClientId(ctx.getClientIdentity().getId());
-        List<String> clientGroups = new GroupContractUtil().getGroupNamesForUser(ctx.getStub(), clientId);
 
         ApprovalList existingApprovals = operationData.getExistingApprovals().addUsersItem(clientId).addGroupsItems(clientGroups);
-        ApprovalList requiredApprovals = null;
+        ApprovalList requiredApprovals;
         try {
             requiredApprovals = AccessManager.getRequiredApprovals(contractName, transactionName, params);
         } catch (MissingTransactionError e) {
@@ -91,7 +93,7 @@ public class OperationContract extends ContractBase {
     public String rejectTransaction(final Context ctx, final String operationId, final String rejectMessage) {
         OperationData operationData;
         try {
-            operationData = cUtil.<OperationData>getState(ctx.getStub(), operationId, OperationData.class);
+            operationData = cUtil.getState(ctx.getStub(), operationId, OperationData.class);
         } catch (LedgerAccessError e) {
             return e.getJsonError();
         }
@@ -105,7 +107,7 @@ public class OperationContract extends ContractBase {
 
         OperationData operationData;
         try {
-            operationData = cUtil.<OperationData>getState(ctx.getStub(), operationId, OperationData.class);
+            operationData = cUtil.getState(ctx.getStub(), operationId, OperationData.class);
         } catch (LedgerAccessError e) {
             return e.getJsonError();
         }